# webapp/openapi/integration-v1.yaml
openapi: 3.1.0
...
info:
  title: Kinboard Integration API
  version: "1.0.0"
  description: |
    Machine-to-machine access to a Kinboard household.

    Authentication is a bearer token created in **Settings → Integrations**.
    Tokens carry explicit scopes; no scope implies another, and every failure
    to authorise answers `401` — including "valid token, wrong scope", so a
    caller probing an endpoint cannot learn that its token is real.
  license:
    name: PolyForm Noncommercial 1.0.0
    identifier: PolyForm-Noncommercial-1.0.0
servers:
  - url: "{kinboard_url}/api/integration/v1"
    variables:
      kinboard_url:
        default: "http://kinboard.local:3000"
        description: The address of the Kinboard instance.
security:
  - integrationToken: []
...
paths:
...
  /calendar/events:
    get:
      tags: [state]
      summary: Calendar events in a range
      description: |
        Backs `calendar.kinboard_family`. Separate from `/family/summary`
        because the two answer different questions: the summary is "what is
        true right now", polled every half minute; a calendar entity is asked
        for arbitrary windows — whatever the user is looking at, possibly next
        month — and returning that on every summary poll would be absurd.

        Selection is by **overlap**, not containment: an event that started
        yesterday and ends tomorrow belongs in today's window. Filtering on
        start alone would drop exactly the long events a calendar most needs
        to show.

        Both bounds are required. A default would be a guess, and the two
        plausible guesses ("today", "everything") differ by orders of
        magnitude in cost.
      x-required-scope: family:read
      parameters:
        - name: start
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: end
          in: query
          required: true
          schema: { type: string, format: date-time }
          description: Must be after `start`, and no more than 370 days later.
      responses:
        "200":
          description: Events overlapping the window, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/CalendarEvent" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "429": { $ref: "#/components/responses/RateLimited" }
    post:
      tags: [state]
      summary: Create a family event with Google or CalDAV write-through
      description: |
        ICS subscriptions and read-only CalDAV calendars cannot be targeted.
        The response reports whether the connected provider accepted the event.

        A **timed** event takes `start_at` and `end_at`, each with `Z` or a
        `+HH:MM` offset (`+0200` is rejected).

        An **all-day** event takes `all_day: true` with `start_date` and
        `end_date` as plain dates, `end_date` being the **last** day
        (inclusive; it defaults to `start_date`). Timestamps are rejected for
        all-day events, because a midnight sent in the wrong zone lands on the
        wrong day. Kinboard converts the dates in the family's time zone.
      x-required-scope: calendar:write
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 255 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [calendar_id, title]
              properties:
                calendar_id: { type: string, format: uuid }
                title: { type: string, maxLength: 300 }
                start_at: { type: string, format: date-time, description: Timed events only. }
                end_at: { type: string, format: date-time, description: Timed events only. }
                all_day: { type: boolean }
                start_date: { type: string, format: date, description: All-day events only. First day. }
                end_date: { type: string, format: date, description: All-day events only. Last day, inclusive. }
                description: { type: string, maxLength: 2000 }
                location: { type: string, maxLength: 300 }
      responses:
        "201":
          description: Created local event
          content:
            application/json:
              schema:
                type: object
                required: [event, sync]
                properties:
                  event:
                    allOf:
                      - { $ref: "#/components/schemas/CalendarEvent" }
                      - type: object
                        required: [calendar_id]
                        properties:
                          calendar_id: { type: string, format: uuid }
                  sync:
                    type: object
                    required: [provider, synced]
                    properties:
                      provider: { type: string, enum: [local, google, caldav] }
                      synced: { type: boolean }
                      reason: { type: string }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { description: Calendar is not writable or in this family. }
        "409": { description: Idempotency key reused with different arguments. }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { description: The event could not be stored. }
...
components:
  securitySchemes:
    integrationToken:
      type: http
      scheme: bearer
      description: A token from Settings → Integrations. Shown once; stored only as a hash.
  responses:
    Unauthenticated:
      description: |
        No token, an unknown/revoked/expired token, or a token without the
        required scope. Deliberately one answer for all four — distinguishing
        them would tell a caller probing an endpoint that its token is real.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidRequest:
      description: The request was understood but its arguments were not usable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: |
        Too many requests for this token. Budgets are per token and per method
        class: reads and writes are counted separately, so an automation that
        writes too often does not also blind a dashboard that only reads.

        `Retry-After` gives the seconds to wait and is never zero.
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable. May be reworded at any time — do not branch on it.
        code:
          type: string
          description: Stable and machine-readable. Branch on this.
          enum:
            - not_authenticated
            - forbidden
            - not_found
            - invalid_request
            - conflict
            - rate_limited
            - upstream_unavailable
            - unavailable
            - internal_error
            - not_implemented
        correlationId:
          type: string
          description: Also sent as the `x-correlation-id` header, and printed on every log line for this request.
    CalendarEvent:
      type: object
      required: [id, title, start_at, end_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: [string, "null"] }
        location: { type: [string, "null"] }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        all_day: { type: [boolean, "null"] }
        person_id: { type: [string, "null"], format: uuid }
...